requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
Django==4.2.2
django-filter==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
openpyxl==3.1.2
phonenumberslite==8.13.14
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pydub==0.25.1
PyJWT==2.7.0
pytz==2023.3
pyVoIP==1.6.4
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
